Wrap in jupyter server
