+++ packages/jupyter-graphql/README.md
@@ -7,6 +7,8 @@ GraphQL powered Jupyter server, that exposes a stateful representation of the ex
 ```bash
 pip install flit
 flit install --symlink
+# https://github.com/jupyter/jupyter_server/tree/master/examples/simple#extension-1
+jupyter server --ServerApp.jpserver_extensions="{'jupyter_graphql': True}"
 uvicorn jupyter_graphql:app --reload
 # Open GraphQL inspector
 open http://127.0.0.1:8000/graphql/
